+++ server/src/image/Image.dto.ts
@@ -1,0 +1,3 @@
+export class ImageDto {
+    [image: string]: string;
+}
